+++ dca/tl_news_archive.php
@@ -16,6 +16,7 @@
  * Register the global callbacks
  */
 $GLOBALS['TL_DCA']['tl_news_archive']['config']['onload_callback'][] = array('tl_news_archive_categories', 'checkPermission');
+$GLOBALS['TL_DCA']['tl_news_archive']['config']['onload_callback'][] = array('tl_news_archive_categories', 'adjustPalette');
 
 /**
  * Add a global operation to tl_news_archive
@@ -31,6 +32,33 @@ array_insert($GLOBALS['TL_DCA']['tl_news_archive']['list']['global_operations'],
     )
 ));
 
+/**
+ * Add palettes to tl_news_archive
+ */
+$GLOBALS['TL_DCA']['tl_news_archive']['palettes']['default'] = str_replace('jumpTo;', 'jumpTo;{categories_legend},limitCategories;', $GLOBALS['TL_DCA']['tl_news_archive']['palettes']['default']);
+
+/**
+ * Add fields to tl_news_archive
+ */
+$GLOBALS['TL_DCA']['tl_news_archive']['fields']['limitCategories'] = array
+(
+    'label'                   => &$GLOBALS['TL_LANG']['tl_news_archive']['limitCategories'],
+    'exclude'                 => true,
+    'inputType'               => 'checkbox',
+    'eval'                    => array('submitOnChange'=>true),
+    'sql'                     => "char(1) NOT NULL default ''"
+);
+
+$GLOBALS['TL_DCA']['tl_news_archive']['fields']['categories'] = array
+(
+    'label'                   => &$GLOBALS['TL_LANG']['tl_news_archive']['categories'],
+    'exclude'                 => true,
+    'inputType'               => 'treePicker',
+    'foreignKey'              => 'tl_news_category.title',
+    'eval'                    => array('mandatory'=>true, 'multiple'=>true, 'fieldType'=>'checkbox', 'foreignTable'=>'tl_news_category', 'titleField'=>'title', 'searchField'=>'title', 'managerHref'=>'do=news&table=tl_news_category'),
+    'sql'                     => "blob NULL"
+);
+
 class tl_news_archive_categories extends Backend
 {
 
@@ -44,5 +72,25 @@ class tl_news_archive_categories extends Backend
         if (!$this->User->isAdmin && !$this->User->newscategories) {
             unset($GLOBALS['TL_DCA']['tl_news_archive']['list']['global_operations']['categories']);
         }
+    }
+
+    /**
+     * Adjust the palette
+     */
+    public function adjustPalette($dc=null)
+    {
+        if (!$dc->id) {
+            return;
+        }
+
+        $objArchive = $this->Database->prepare("SELECT limitCategories FROM tl_news_archive WHERE id=?")
+                                     ->limit(1)
+                                     ->execute($dc->id);
+
+        if (!$objArchive->numRows || !$objArchive->limitCategories) {
+            return;
+        }
+
+        $GLOBALS['TL_DCA']['tl_news_archive']['palettes']['default'] = str_replace('limitCategories;', 'limitCategories,categories;', $GLOBALS['TL_DCA']['tl_news_archive']['palettes']['default']);
     }
 }